CIRC-2228: TLRs in Search slips missing key tokens
