+++ src/main/java/org/folio/circulation/domain/notice/combiner/LoanNoticeContextCombiner.java
@@ -1,12 +1,10 @@
 package org.folio.circulation.domain.notice.combiner;
 
-import static java.util.stream.Collectors.collectingAndThen;
-import static org.folio.circulation.domain.notice.TemplateContextUtil.createUserContext;
-
 import java.util.Collection;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
+import org.folio.circulation.domain.mapper.UserMapper;
 import org.folio.circulation.domain.notice.PatronNoticeEvent;
 import org.folio.circulation.domain.representations.logs.NoticeLogContext;
 
@@ -19,10 +17,10 @@ public class LoanNoticeContextCombiner implements NoticeContextCombiner {
     return events.stream()
       .map(PatronNoticeEvent::getNoticeContext)
       .filter(Objects::nonNull)
-      .collect(collectingAndThen(
+      .collect(Collectors.collectingAndThen(
         Collectors.toList(),
         contexts -> new JsonObject()
-          .put("user", createUserContext(events.iterator().next().getUser()).asJson())
+          .put("user", UserMapper.createUserContext(events.iterator().next().getUser()))
           .put("loans", contexts)
       ));
   }
@@ -35,7 +33,7 @@ public class LoanNoticeContextCombiner implements NoticeContextCombiner {
       .map(NoticeLogContext::getItems)
       .flatMap(Collection::stream)
       .filter(Objects::nonNull)
-      .collect(collectingAndThen(
+      .collect(Collectors.collectingAndThen(
         Collectors.toList(),
         items -> new NoticeLogContext()
           .withUser(events.iterator().next().getUser())
